fix: correct Decimal comparisons and parseInt truncation in payment paths
- webhooks.ts handleSetupIntentSuccess: wrap currentProjectAmount and Decimal
  fields with Number() before >= and - comparisons (Decimal.valueOf() returns
  a string so Decimal >= Decimal was doing lexicographic string comparison,
  causing incorrect funding-goal detection)
- webhooks.ts: pass Number(existingPledge.amount) to notifyPledgeReceived
  which expects a number, not a Prisma Decimal
- RewardSelector.tsx: change parseInt → parseFloat on custom pledge amount
  input so $9.99 is not silently truncated to $9

+++ src/app/projects/[vanityname]/[slug]/pledge/components/RewardSelector.tsx
@@ -69,7 +69,7 @@ export function RewardSelector({
                   type="number"
                   min={1}
                   value={customPledgeAmount}
-                  onChange={(e) => setCustomPledgeAmount(Math.max(1, parseInt(e.target.value) || 1))}
+                  onChange={(e) => setCustomPledgeAmount(Math.max(1, parseFloat(e.target.value) || 1))}
                   className="w-20 border-0 focus-visible:ring-0 bg-transparent"
                 />
               </div>
